### 2024-06-26 e312118
Add a JOSS badge
