### 2024-06-26 #6931 @AkchurinDA
Hi,

I would like to suggest the following potential reviewers for the paper based on their respective research domains listed [here](https://reviewers.joss.theoj.org/lookup):
- baxmittens - uncertainty quantification
- rafaelorozco - uncertainty quantification

I would also be happy to provide a list of potential reviewers outside of that list.

Thank you.
