### 2024-06-27 #6931 @jbytecode
👋👋👋 Dear @baxmittens, @rafaelorozco 👋👋👋

[[PRE REVIEW]: Fortuna.jl: Structural and System Reliability Analysis in Julia](https://github.com/openjournals/joss-reviews/issues/6931#top)

You can find more information at the top of this Github issue (Provided above link).

This is the pre-review issue. After setting at least 2 reviewers we will start the review process in a separate thread. In that thread, there will be nearly 25 check items for each single reviewer.

Thank you in advance!

@rafaelorozco - I know you have one more workload at JOSS, but I think you can sequentially handle them. Please tell me otherwise.
