### 2024-06-28 #6931 @AkchurinDA
Hi, @jbytecode,

Thanks for agreeing to be the editor!

I have a quick question: It is OK if I update the paper (fixed a few typos and improved the readability of a few sentences) before the reviewers get assigned?
