### 2024-06-28 a1edd89
Update the JOSS paper
