+++ JOSS/paper.md
@@ -8,6 +8,7 @@ authors:
     - name: Damir Akchurin
       orcid: 0000-0003-0849-4917
       affiliation: 1
+      corresponding: true
 affiliations:
     - name: Department of Civil and Systems Engineering, Johns Hopkins University, Baltimore, MD
       index: 1
@@ -17,7 +18,7 @@ bibliography: paper.bib
 
 # Summary
 
-In any field of engineering, addressing uncertainties is crucial to ensure the adequate reliability – ability to function without failure – of structures and systems, such as residential and commercial buildings, bridges, nuclear power plants, energy distribution networks, transportation networks, and even electric circuits. In these systems, uncertainties are ubiquitous and present in their geometric and material properties, the demands acting on them, and the measurements used to assess their condition. Moreover, uncertainties can also be present in the probabilistic models used to quantify them. As a result of the irreducibility of some of these uncertainties, it is impossible to guarantee the absolute reliability and continuous serviceability of a system at any given point in time with certainty. However, it is possible to define and quantify measures of reliability, such as failure probability of a system under given demands, using probabilistic methods developed within the field of structural and system reliability.
+In any field of engineering, addressing uncertainties is crucial to ensure the adequate reliability – ability to function without failure – of structures and systems, such as residential and commercial buildings, bridges, nuclear power plants, energy distribution networks, transportation networks, and even electric circuits. In these systems, uncertainties are ubiquitous and present in their geometric and material properties, the demands acting on them, and the measurements used to assess their condition. Moreover, uncertainties can also be present in the probabilistic models used to quantify them. As a result of the irreducibility of some of these uncertainties, it is impossible to guarantee the absolute reliability and continuous serviceability of a system at any given point in time with certainty. However, it is possible to define and quantify measures of reliability, such as the failure probability of a system under given demands, using probabilistic methods developed within the field of structural and system reliability.
 
 `Fortuna.jl` is an open-source general-purpose package for structural and system reliability analysis purely written in the Julia programming language [@Bezanson:2017]. It implements a wide suite of commonly used reliability analysis methods to solve reliability, inverse reliability, and sensitivity problems. `Fortuna.jl` leverages Julia’s high-performance capabilities to efficiently solve even the most challenging reliability problems. At the same time, `Fortuna.jl` is designed to be user-friendly and flexible, making it suitable for both research and teaching settings. It is intended that `Fortuna.jl` can serve as a platform for the development and implementation of new rapidly emerging reliability analysis methods.
 
@@ -27,7 +28,7 @@ Due to largely numerical nature of almost all reliability analysis methods, whic
 
 The development of `Fortuna.jl` was mainly motivated by the absence of packages for structural and system reliability analysis written in Julia. Additionally, it aimed to achieve an improved balance between user experience, ease of implementing new reliability analysis methods, computational efficiency, and interoperability with external finite element (FE) modeling software not directly available through Julia.
 
[…]
