+++ JOSS/paper.bib
@@ -235,3 +235,15 @@
 	year = {2013},
 	pages = {232--240},
 }
+
+@misc{Surrogates:2024,
+	title = {{SciML}/{Surrogates}.jl: {V6}.10.0},
+	copyright = {Creative Commons Attribution 4.0 International},
+	shorttitle = {{SciML}/{Surrogates}.jl},
+	url = {https://zenodo.org/doi/10.5281/zenodo.12571719},
+	urldate = {2024-06-28},
+	publisher = {Zenodo},
+	month = jun,
+	year = {2024},
+	doi = {10.5281/ZENODO.12571719},
+}
