### 2024-06-28 6a8b362
Update the colors in the figure and add a reference to Makie
