### 2024-06-28 #6931 @AkchurinDA
Thanks, @jbytecode!

Sure, here is a list of 3 potential external reviewers who work in the field of structural and system reliability analysis and with whom I don't have any conflicts of interest:
- [Dr. Brennan Bean](https://www.usu.edu/math/directory/faculty/bean-brennan)
- [Dr. Abbie B. Liel](https://www.colorado.edu/ceae/abbie-b-liel)
- [Dr. Jared DeBock](https://www.csuchico.edu/ce/people/faculty/jared-debock.shtml)
