### 2024-06-28 #6931 @AkchurinDA
@jbytecode

Oh, sorry, I think I misunderstood. These people are outside of the JOSS' list of reviewers and I don't know their GitHub handles.

I would propose the following potential reviewers from the list:
- [dannys4](https://github.com/dannys4) - uncertainty quantification, numerical methods
- [ziyiyin97](https://github.com/ziyiyin97) - uncertainty quantification,  numerical methods
- [dmbates](https://github.com/dmbates) - statistics
- [ahwillia](https://github.com/ahwillia) - statistics
