### 2024-06-28 #6931 @jbytecode
@danielskatz - Thank you very much for the explanation. I had already mentioned that we only need GitHub account names. The list keeps the information of those who want to review for JOSS, but of course, we can invite other GitHub accounts from outside as well.
