### 2024-06-28 #6931 @danielskatz
Editors can also search for GitHub accounts of potential reviewers, or contact them via email, where they may either say that they have a GitHub account already or are willing to create one to do the review.
