### 2024-06-30 #6931 @jbytecode
Dear Dr. Brennan Bean (@beanb2)

I'm writing to you because you have been suggested as a potential reviewer for a paper in JOSS (Journal of Open Source Software).

Would you be willing to assist in reviewing this submission for JOSS?

**Fortuna.jl: Structural and System Reliability Analysis in Julia** (https://github.com/openjournals/joss-reviews/issues/6931)

You can find more information at the top of this Github issue (Provided above link).

If you accept reviewing, you can post a message in the GitHub link https://github.com/openjournals/joss-reviews/issues/6931

Thank you in advance!
