### 2024-06-30 #6931 @beanb2
Thank you for the offer. I would normally be very interested in this kind of opportunity and hope that JOSS will keep me in mind for future reviews. That said, I my wife is recovering from cancer treatment and I am unable to take on any additional responsibilities for the next 6 weeks. Assuming you want a timely review, I will need to pass on this offer.

Brennan

From: Mehmet Hakan Satman ***@***.***>
Sent: Sunday, June 30, 2024 11:08 AM
Cc: Brennan Bean ***@***.***>; Mention ***@***.***>
Subject: Re: [openjournals/joss-reviews] [PRE REVIEW]: Fortuna.jl: Structural and System Reliability Analysis in Julia (Issue #6931)

Dear Dr. Brennan Bean ***@***.***<https://github.com/beanb2>)

I'm writing to you because you have been suggested as a potential reviewer for a paper in JOSS (Journal of Open Source Software).

Would you be willing to assist in reviewing this submission for JOSS?

Fortuna.jl: Structural and System Reliability Analysis in Julia (#6931<https://github.com/openjournals/joss-reviews/issues/6931>)

You can find more information at the top of this Github issue (Provided above link).

If you accept reviewing, you can post a message in the GitHub link #6931<https://github.com/openjournals/joss-reviews/issues/6931>

Thank you in advance!
