### 2024-07-02 #6931 @AkchurinDA
@beanb2 - Sorry to hear that, Dr. Bean. I hope your wife recovers as soon as possible. I will keep you in my thoughts.
