### 2024-07-03 #6931 @jbytecode
👋👋👋 Dear @dannys4  👋👋👋

[[PRE REVIEW]: Fortuna.jl: Structural and System Reliability Analysis in Julia](https://github.com/openjournals/joss-reviews/issues/6931#top)

You can find more information at the top of this Github issue (Provided above link).

This is the pre-review issue. After setting at least 2 reviewers we will start the review process in a separate thread. In that thread, there will be nearly 25 check items for each single reviewer.

Thank you in advance!
