### 2024-07-07 29f95b0
Update JOSS paper
