### 2024-07-07 9cd70f7
Update the color palette
