+++ JOSS/paper.md
@@ -18,7 +18,7 @@ bibliography: paper.bib
 
 # Summary
 
-In any field of engineering, addressing uncertainties is crucial to ensure the adequate reliability – ability to function without failure – of structures and systems, such as residential and commercial buildings, bridges, nuclear power plants, energy distribution networks, transportation networks, and even electric circuits. In these systems, uncertainties are ubiquitous and present in their geometric and material properties, the demands acting on them, and the measurements used to assess their condition. Moreover, uncertainties can also be present in the probabilistic models used to quantify them. As a result of the irreducibility of some of these uncertainties, it is impossible to guarantee the absolute reliability and continuous serviceability of a system at any given point in time with certainty. However, it is possible to define and quantify measures of reliability, such as the failure probability of a system under given demands, using probabilistic methods developed within the field of structural and system reliability.
+In any field of engineering, addressing uncertainties is crucial to ensure the adequate reliability – ability to function without failure – of structures and systems, such as residential and commercial buildings, bridges, nuclear power plants, energy distribution networks, transportation networks, and even electric circuits. In these systems, uncertainties are ubiquitous and present in their geometric and material properties, the demands acting on them, and the measurements used to assess their condition. Moreover, uncertainties can also be present in the probabilistic models used to quantify them. As a result of the irreducibility of some of these uncertainties, it is impossible to guarantee the absolute reliability and serviceability of a system at any given point in time with certainty. However, it is possible to define and quantify measures of reliability, such as the failure probability of a system under given demands, using probabilistic methods developed within the field of structural and system reliability.
 
 `Fortuna.jl` is an open-source general-purpose package for structural and system reliability analysis purely written in the Julia programming language [@Bezanson:2017]. It implements a wide suite of commonly used reliability analysis methods to solve reliability, inverse reliability, and sensitivity problems. `Fortuna.jl` leverages Julia’s high-performance capabilities to efficiently solve even the most challenging reliability problems. At the same time, `Fortuna.jl` is designed to be user-friendly and flexible, making it suitable for both research and teaching settings. It is intended that `Fortuna.jl` can serve as a platform for the development and implementation of new rapidly emerging reliability analysis methods.
 
