### 2024-07-07 1273f99
Resize the figure
