### 2024-07-08 #6967 @editorialbot
[REVIEW]: Fortuna.jl: Structural and System Reliability Analysis in Julia
**Submitting author:** @AkchurinDA (<a href="http://orcid.org/0000-0003-0849-4917">Damir Akchurin</a>)
**Repository:** https://github.com/AkchurinDA/Fortuna.jl
**Branch with paper.md** (empty if default branch): JOSS
**Version:** v0.10.1
