+++ JOSS/paper.md
@@ -18,19 +18,19 @@ bibliography: paper.bib
 
 # Summary
 
-In any field of engineering, addressing uncertainties is crucial to ensure the adequate reliability – ability to function without failure – of structures and systems, such as residential and commercial buildings, bridges, nuclear power plants, energy distribution networks, transportation networks, and even electric circuits. In these systems, uncertainties are ubiquitous and present in their geometric and material properties, the demands acting on them, and the measurements used to assess their condition. Moreover, uncertainties can also be present in the probabilistic models used to quantify them. As a result of the irreducibility of some of these uncertainties, it is impossible to guarantee the absolute reliability and serviceability of a system at any given point in time with certainty. However, it is possible to define and quantify measures of reliability, such as the failure probability of a system under given demands, using probabilistic methods developed within the field of structural and system reliability.
+In any engineering field, addressing uncertainties is crucial to ensure the adequate reliability – the ability to function without failure – of structures and systems, such as residential and commercial buildings, bridges, nuclear power plants, energy distribution networks, transportation networks, and even electric circuits. In these systems, uncertainties are ubiquitous and present in their geometric and material properties, the demands acting on them, and the measurements used to assess their condition. Moreover, uncertainties can also be present in the probabilistic models used to quantify them. As a result of the irreducibility of some of these uncertainties, it is impossible to guarantee a system's absolute reliability and serviceability at any given point in time with certainty. However, it is possible to define and quantify measures of reliability, such as the failure probability of a system under given demands, using probabilistic methods developed within the field of structural and system reliability.
 
-`Fortuna.jl` is an open-source general-purpose package for structural and system reliability analysis purely written in the Julia programming language [@Bezanson:2017]. It implements a wide suite of commonly used reliability analysis methods to solve reliability, inverse reliability, and sensitivity problems. `Fortuna.jl` leverages Julia’s high-performance capabilities to efficiently solve even the most challenging reliability problems. At the same time, `Fortuna.jl` is designed to be user-friendly and flexible, making it suitable for both research and teaching settings. It is intended that `Fortuna.jl` can serve as a platform for the development and implementation of new rapidly emerging reliability analysis methods.
+`Fortuna.jl` is an open-source general-purpose package for structural and system reliability analysis purely written in the Julia programming language [@Bezanson:2017]. It implements a wide suite of commonly used reliability analysis methods to solve reliability, inverse reliability, and sensitivity problems. `Fortuna.jl` leverages Julia’s high-performance capabilities to solve even the most challenging reliability problems efficiently. At the same time, `Fortuna.jl` is designed to be user-friendly and flexible, making it suitable for both research and teaching settings. It is intended that `Fortuna.jl` can serve as a platform for developing and implementing new rapidly emerging reliability analysis methods.
 
 # Statement of Need
 
-Due to largely numerical nature of almost all reliability analysis methods, which often require computation of gradient vectors and Hessian matrices, the use of computational resources is necessary for all but the most trivial problems. As a result, a large number of both open-source and commercial software packages have been developed in different programming languages over the past 30 years, such as `UQpy` (Python) [@Olivier:2020], `Pystra` (Python), `FERUM` (MATLAB) [@Bourinet:2009], and `CalREL` (FORTRAN) [@DerKiureghian:2006].
[…]
