### 2024-07-09 #6967 @baxmittens
Hi @AkchurinDA

I ran Grammarly over your paper and created a pull request.

https://github.com/AkchurinDA/Fortuna.jl/pull/85

Double-check carefully before accepting the changes or rejecting them as you like.

I will continue to review it later this week. From what I have seen, it looks like an exciting package.
