### 2024-07-22 #6967 @rafaelorozco
I have no such justification, I have been doing a couple checkmarks a day. I hope that my pace is acceptable!

I was thinking that community guidelines specifically for how to contribute to the software would be nice. I see you have a section in the readme for raising issue so maybe just add contribute guidelines there to make that clear.
