### 2024-07-22 #6967 @AkchurinDA
Hi @rafaelorozco,
Thank you for the suggestion. I will add a simple contribution guideline in a few hours.
