### 2024-07-22 #6967 @AkchurinDA
@rafaelorozco,

I have added a simple contribution guideline [here](https://github.com/AkchurinDA/Fortuna.jl/blob/main/CONTRIBUTING.md).
