+++ JOSS/paper.md
@@ -30,11 +30,7 @@ The development of `Fortuna.jl` was mainly motivated by the absence of packages 
 
 A key distinguishing feature of `Fortuna.jl` is that it is capable of performing differentiation of the limit state function, which defines the failure criterion for a given reliability problem, required by most reliability analysis methods using automatic differentiation techniques provided by `ForwardDiff.jl` package [@Revels:2016], significantly speeding up the analysis process. This contrasts with other software packages for structural and system reliability analysis, which typically rely on variants of finite difference approximations. As it was mentioned before, `Fortuna.jl` also allows to easily define limit state functions using external FE modeling software, such as `Abaqus`, `OpenSees` [@McKenna:2010], and `SAFIR` [@Franssen:2017]. `Fortuna.jl` can also easily work with surrogate models of limit state functions that are computationally expensive to evaluate, such as when a limit state function is defined in terms of a complex FE model, using the functionality provided by `Surrogates.jl` package [@Surrogates:2024]. Lastly, `Fortuna.jl`’s ability to work with random variables is based on a widely adopted `Distributions.jl` package [@Besancon:2021], enabling seamless integration with other software packages for probabilistic analysis written in Julia, such as `Copulas.jl` [@Laverny:2024], `RxInter.jl` [@Bagaev:2023], and `Turing.jl` [@Ge:2018], which provide all the functionality needed for the development and implementation of more modern reliability analysis methods based on Bayesian inference.
 
-<<<<<<< HEAD
-`Fortuna.jl` has already been successfully used to compare the levels of safety achieved by the Allowable Strength Design and Load (ASD) and Resistance Factor Design (LRFD) methods, two design frameworks used to design individual component in structures [@Sabelli:2024; @Akchurin:2024]. `Fortuna.jl` is also being actively used to develop and implement a next-generation design framework that can account for and optimize system behavior in the design of steel structures, which cannot be achieved using current methods such as the ASD and LRFD.
-=======
 `Fortuna.jl` has already been successfully used to compare the levels of safety achieved by the Allowable Strength Design (ASD) and Load and Resistance Factor Design (LRFD) methods, two design frameworks used to design individual component in structures [@Sabelli:2024; @Akchurin:2024]. `Fortuna.jl` is also being actively used to develop and implement a next-generation design framework that can account for and optimize system behavior in the design of steel structures, which cannot be achieved using the current methods such as the ASD and LRFD.
->>>>>>> fd1b642 (Fix a typo)
 
 # Example
 
