### 2024-07-26 #6967 @AkchurinDA
Hi @jbytecode,

Hope you are doing well!

I wanted to ask a quick question. I want to push out a small non-breaking update to the package, is it somehow possible to update the version of the package referenced in the submission?
