### 2024-07-26 #6967 @baxmittens
> Hi @jbytecode,
>
> Hope you are doing well!
>
> I wanted to ask a quick question. I want to push out a small non-breaking update to the package, is it somehow possible to update the version of the package referenced in the submission?
>
> Thank you!

As far as I know, you can do this without a problem.
You have to create a Zenodo archive of the most recent version of your package after the paper is accepted. This archive is then linked with the paper together with the GitHub page of the package.
