### 2024-07-26 #6967 @baxmittens
@AkchurinDA
Can you add authors to your citation of SciML/Surrogates.jl?
According to the Zenodo archive, it should be something like Bessi, L. et. al.
