### 2024-07-26 a7cbb62
Update the reference for Surrogates.jl
