+++ JOSS/paper.bib
@@ -237,13 +237,14 @@
 }
 
 @misc{Surrogates:2024,
-	title = {{SciML}/{Surrogates}.jl: {V6}.10.0},
+	title = {{SciML}/{Surrogates}.jl: v6.10.0},
 	copyright = {Creative Commons Attribution 4.0 International},
 	shorttitle = {{SciML}/{Surrogates}.jl},
-	url = {https://zenodo.org/doi/10.5281/zenodo.12571719},
-	urldate = {2024-06-28},
+	url = {https://zenodo.org/doi/10.5281/zenodo.12571718},
+	urldate = {2024-07-26},
 	publisher = {Zenodo},
+	author = {Bessi, Ludovico and Rackauckas, Christopher and , Vikram and Singh Rathaur, Rohit and Bhagavan, Sathvik and Marks, Thomas and Anantharaman, Ranjan and , marcoq and Strouwen, Arno and Cognolato, Andrea and Foiles, Dreycen and Latawiec, Pawel and Bharambe, Ashutosh and , michiboo and Yalburgi, Sharan and Baker, Fergus and Fuhrmann, Jürgen and , st-- and Thazhemadam, Anant and de Graaf, Chris and , jbrea and Shah, Viral B. and Bhogaonker, Krishna and Haldar, Anubhab and Piibeleht, Morten and Martinuzzi, Francesco and Gupta, Kanav and Sarnoff, Jeffrey},
 	month = jun,
 	year = {2024},
-	doi = {10.5281/ZENODO.12571719},
+	doi = {10.5281/ZENODO.12571718},
 }
