### 2024-07-26 #6967 @rafaelorozco
Nothing on my side to speed up, i have been slowly going through the checklist. I just finished the checklist. I can send a short report soon but I can already say that I am satisfied with the current state of the package/writeup and recommend them for publication in JOSS.
