### 2024-07-26 #6967 @AkchurinDA
@baxmittens -  I have updated the author list for the reference on `Surrogate.jl` package. Thanks for catching it!

@rafaelorozco - Thank you very much for going over the checklist so quickly, I appreciate it!
