### 2024-07-26 b3810ce
Log the bug fix into NEWS.md
