### 2024-07-26 #6967 @AkchurinDA
@jbytecode - here is [link](https://github.com/AkchurinDA/Fortuna.jl/pull/87/commits) to the PR that fixed a problem I had with the Latin Hypercube Sampling technique, which was failing for vey large number of samples (>1E8). I have triggered the JuliaRegiatrator bot, [new version (V0.10.1) should be up soon](https://github.com/JuliaRegistries/General/pull/111846).
