+++ v0.10.1/search_index.js
@@ -1,0 +1,3 @@
+var documenterSearchIndex = {"docs":
+[{"location":"Reliability Problems/MC/#Monte-Carlo-Simulations","page":"Direct Monte Carlo Simulations","title":"Monte Carlo Simulations","text":"","category":"section"},{"location":"Reliability Problems/MC/#Overview","page":"Direct Monte Carlo Simulations","title":"Overview","text":"","category":"section"},{"location":"Reliability Problems/MC/","page":"Direct Monte Carlo Simulations","title":"Direct Monte Carlo Simulations","text":"Using direct Monte Carlo simulations (MCS) is the simplest way to estimate the probability of failure P_f for reliability problems with both simple and complex limit state functions g(vecX). The brute force MCS relies on the following reformulation of the general analytical expression for the probability of failure P_f:","category":"page"},{"location":"Reliability Problems/MC/","page":"Direct Monte Carlo Simulations","title":"Direct Monte Carlo Simulations","text":"P_f = P(Omega_f) = int_Omega_f f_vecX(vecx) dvecx = int_mathbbR^n mathbbI(vecx) f_vecX(vecx) dvecx = mathbbE_fmathbbI(vecx)","category":"page"},{"location":"Reliability Problems/MC/","page":"Direct Monte Carlo Simulations","title":"Direct Monte Carlo Simulations","text":"where f_vecX(vecx) is the joint probability function of the input random vector vecX, Omega_f = vecX g(vecX) leq 0 is the failure domain defined by the limit state function g(vecX), and mathbbI(vecx) is the indicator function given by:","category":"page"},{"location":"Reliability Problems/MC/","page":"Direct Monte Carlo Simulations","title":"Direct Monte Carlo Simulations","text":"mathbbI(vecx) = \nbegincases\n    1  textif  vecx in Omega_f \n    0  textotherwise\nendcases","category":"page"},{"location":"Reliability Problems/MC/","page":"Direct Monte Carlo Simulations","title":"Direct Monte Carlo Simulations","text":"Therefore, the probability of failure P_f is defined as the expectation of the indicator function mathbbI(vecx). If samples of the input random vector vecx are generated numerically, then the estimator of the probability of failure hatP_f is","category":"page"},{"location":"Reliability Problems/MC/","page":"Direct Monte Carlo Simulations","title":"Direct Monte Carlo Simulations","text":"hatP_f = dfrac1N sum_i = 1^N mathbbI(vecx_i)","category":"page"},{"location":"Reliability Problems/MC/","page":"Direct Monte Carlo Simulations","title":"Direct Monte Carlo Simulations","text":"where N is the number of generated sampled. The estimator hatP_f is unbiased, i.e., it correctly predicts the true probability of failure, such that  mathbbE(hatP_f) = P_f. The main drawback of using the MCS is that is becomes prohibitively expensive to use if the true probability is too small, e.g., P_f leq 10^-6, given that the variance of the estimator is inversely proportional to the number of generated samples, such that,","category":"page"},{"location":"Reliability Problems/MC/","page":"Direct Monte Carlo Simulations","title":"Direct Monte Carlo Simulations","text":"textVar(hatP_f) = frac1N P_f (1 - P_f)","category":"page"},{"locatio
[…]
