+++ v0.10.1/References.bib
@@ -1,0 +1,134 @@
+@techreport{Denavit:2013,
+  author = {Denavit, Mark D. and Hajjar, Jerome F.},
+  title = {{Description} of {Geometric} {Nonlinearity} for {Beam}-{Column} {Analysis} in {OpenSees}},
+  number = {NEU-CEE-2013-02},
+  institution = {Department of Civil and Environmental Engineering, Northeastern University},
+  year = {2013},
+}
+
+@article{Akchurin:2024,
+	title = {{ASD} and {LRFD}: {Reliability} {Comparison} for {Designs} {Subjected} to {Wind} {Loads}},
+	author = {Akchurin, Damir and Sabelli, Rafael and Ziemian, Ronald D. and Schafer, Benjamin W.},
+	journal = {Journal of Constructional Steel Research},
+	doi = {10.1016/j.jcsr.2023.108327},
+	volume = {213},
+	pages = {108327},
+	month = feb,
+	year = {2024},
+}
+
+@article{Rosenblatt:1952,
+	title = {{Remarks} on a {Multivariate} {Transformation}},
+	author = {Rosenblatt, Murray},
+	journal = {The Annals of Mathematical Statistics},
+	volume = {23},
+	number = {3},
+	pages = {470--472},
+	month = sep,
+	year = {1952},
+}
+
+@article{Nataf:1962,
+	title = {{Détermination} des {Distribution} dont les {Marges} sont {Données}},
+	author = {Nataf, André},
+	journal = {Comptes Rendus de l'Académie des Sciences},
+	volume = {225},
+	pages = {42--43},
+	year = {1962},
+}
+
+@article{Lebrun:2009,
+	title = {{An} {Innovating} {Analysis} of the {Nataf} {Transformation} from the {Copula} {Viewpoint}},
+	author = {Lebrun, Régis and Dutfoy, Anne},
+	journal = {Probabilistic Engineering Mechanics},
+	doi = {10.1016/j.probengmech.2008.08.001},
+	issn = {02668920},
+	volume = {24},
+	number = {3},
+	pages = {312--320},
+	month = jul,
+	year = {2009},
+	
+}
+
+@book{DerKiureghian:2022,
+	title = {{Structural} and {System} {Reliability}},
+	author = {Der Kiureghian, Armen},
+	publisher = {Cambridge University Press},
+	isbn = {978-1-108-99188-9},
+	year = {2022},
+}
+
+@article{Tvedt:1990,
+	title = {{Distribution} of {Quadratic} {Forms} in {Normal} {Space}—{Application} to {Structural} {Reliability}},
+	author = {Tvedt, Lars},
+	journal = {Journal of Engineering Mechanics},
+	doi = {10.1061/(ASCE)0733-9399(1990)116:6(1183)},
+	issn = {0733-9399, 1943-7889},
+	volume = {116},
+	number = {6},
+	pages = {1183--1197},
+	month = jun,
+	year = {1990},
+}
+
+@article{Hohenbichler:1988,
+	title = {{Improvement} Of {Second}‐{Order} {Reliability} {Estimates} by {Importance} {Sampling}},
+	author = {Hohenbichler, Michael and Rackwitz, Ruediger},
+	journal = {Journal of Engineering Mechanics},
+	doi = {10.1061/(ASCE)0733-9399(1988)114:12(2195)},
+	issn = {0733-9399, 1943-7889},
+	volume = {114},
+	number = {12},
+	pages = {2195--2199},
+	month = dec,
+	year = {1988},
+}
+
+@article{Breitung:1984,
+	title = {{Asymptotic} {Approximations} for {Multinormal} {Integrals}},
+	author = {Breitung, Karl},
[…]
