### 2024-07-29 #6967 @baxmittens
@AkchurinDA
One thing I noticed in the documentation concerns the nomenclature.
You create distributions from their moments...so far, so good, but why is the second moment the standard deviation and not the variance?
![image](https://github.com/user-attachments/assets/d85a1e5e-b89d-427e-a580-94cc016dcf2a)
