### 2024-07-29 #6967 @AkchurinDA
Hi @baxmittens,

This is a great question. This is actually done by design because all of the literature in the field of Structural and System Reliability Analysis typically report the statistic of random variables in terms of the mean $\mu$ and coefficient of variation $V = \sigma / \mu$ values. Here are just a few examples:
- https://ascelibrary.org/doi/full/10.1061/%28ASCE%29ST.1943-541X.0002011
- https://www.sciencedirect.com/science/article/pii/S0143974X23004376
- https://www.sciencedirect.com/science/article/pii/S0143974X23005540
- https://www.sciencedirect.com/science/article/pii/0167473082900121

Hope this answers your question.
