### 2024-07-30 #6967 @AkchurinDA
HI @baxmittens,

Hope everything is well.

Here is a link to a [Dropbox folder](https://www.dropbox.com/scl/fo/e41j15dlf9fcxmdx55l9i/AAg2tFRUTRLkbEY47dUCrNE?rlkey=fo7cr2353km54ka6v8enuyje4&dl=0) with the PDFs of the papers I have mentioned earlier.

Please let me know if there is anything else I can do to help speed up the review.

Thanks!
