### 2024-08-04 #6967 @AkchurinDA
HI @jbytecode,

I believe that @rafaelorozco has already recommended the package for publication.

> Nothing on my side to speed up, i have been slowly going through the checklist. I just finished the checklist. I can send a short report soon but I can already say that I am satisfied with the current state of the package/writeup and recommend them for publication in JOSS.
