### 2024-08-05 #6967 @AkchurinDA
@rafaelorozco @baxmittens - thanks you for recommending the paper for publication! Appreciate you spending the time on the review.

@jbytecode - should I create a Zenodo archive of the latest version of the package?
