### 2024-08-05 #6967 @jbytecode
Note for missing DOIs: I have performed a Crossref search but failed to find DOI's for those two citations
