### 2024-08-05 #6967 @jbytecode
@AkchurinDA - Now, I will check the manuscript and the bibliography in a great detail. Meanwhile I am proof-reading please

- Create a tagged release of the software with a version number in form of `vx.y.z`, e.g. `v0.10.2`, with a caption describing it's belong to a JOSS submission
- Please use that tagged release of the software, create a Zenodo archive. The archive should have the same metadata of off the submission, e.g. the title of the submisson, the list of authors, the ORCID's etc.
- Please report me the tagged release, the link of the archive, and the DOI provided by the Zenodo.

Thank you in advance.
