### 2024-08-05 #6967 @AkchurinDA
@jbytecode,

I have created a Zenodo archive based off the latest release I have (v0.10.1). All of the archive's metadata is based on the submission. Here is all the information you have requested:
- Tagged release: [v0.10.1](https://github.com/AkchurinDA/Fortuna.jl/releases/tag/v0.10.1)
- Link to the archive: https://zenodo.org/records/13222060
- DOI provided by Zenodo: https://doi.org/10.5281/zenodo.13222060

Please let me know if there is anything else I should do.

Thanks again!
