### 2024-08-05 #6967 @jbytecode
@AkchurinDA - I've just sent a pull request (https://github.com/AkchurinDA/Fortuna.jl/pull/88). This PR includes minor fixes on the text. Please review the changes. If you're agreed with them, please merge. Thank you in advance.
