### 2024-08-05 #6967 @AkchurinDA
@jbytecode - Thank you for suggestions, I have just merged the PR. Let us quickly check the paper one more time.
