### 2024-08-05 #6967 @jbytecode
@AkchurinDA - The manuscript and the bibliography seem okay to me. Please take a look at the paper and have a proof read one more time. When you've done with it, please ping me so we can go forward.
