### 2024-08-05 #6967 @AkchurinDA
Yes, I understand what you mean, but based on all the literature on UQ I have read, there is no need to capitalize the word "normal" in this particular use case.

OK, I will go through the paper one more time and let you know. Thanks!
