### 2024-08-05 #6967 @jbytecode
@AkchurinDA - JOSS pdf compiler does not work well with `dfrac` instruction of LaTeX. Could you please merge the PR (https://github.com/AkchurinDA/Fortuna.jl/pull/89) that I've just sent? Thank you in advance.
