### 2024-08-05 2043b97
Change type of the input of the limit state function to AbstractVector
