### 2024-08-05 #6967 @AkchurinDA
@jbytecode - Thanks for the PR. I have merged it.

I have also decided to change the type of the input for the limit state function $g(\boldsymbol{x})$ in my example to `AbstractVector`. This is more general and better approach if the reader will use this example as a starting point for using the package. Hope that's OK.
