### 2024-08-05 #6967 @jbytecode
@AkchurinDA - I agree with the change, `AbstractVector` will handle some other type of data types inherits the AbstractVector now.
