### 2024-08-05 #6967 @danielskatz
@AkchurinDA - I'm suggesting changes in your paper in https://github.com/AkchurinDA/Fortuna.jl/pull/90 - please merge this, or let me know what you disagree with, then we can continue.
