### 2024-08-05 #6967 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
