### 2024-08-05 #6967 @danielskatz
Congratulations to @AkchurinDA (Damir Akchurin) on your publication!!

JOSS relies on volunteers like you to make this process of recognizing and improving open source software work
