### 2024-08-05 #6967 @AkchurinDA
Thank you all for putting in time and effort into reviewing and editing the submission, @danielskatz @jbytecode @baxmittens @rafaelorozco! I really appreciate it!

The review process in the Journal of Open Source Software is just great - it was such a great experience compared to more conventional journals!
