### 2024-08-05 #6967 @AkchurinDA
@danielskatz - On the off topic. The paper preview is not showing on the journal's [website](https://joss.theoj.org/papers/10.21105/joss.06967). I have tried Safari and Chrome, but without success. Does it just need time to process?
