### 2024-08-05 #6967 @danielskatz
Do you mean this?

Or something else? If this, and if you didn't follow the instructions when the paper was accepted `Wait five minutes, then verify that the paper DOI resolves` you might have gotten a version that didn't resolve, which can be cached for a while.
