### 2024-08-05 #6967 @AkchurinDA
Yes, but mine looks like this for some reason in both browsers...

I didn't use the DOI provided by the `editorialbot`, I just went on the JOSS' website and found the paper. It should be a problem on my side then, just wanted to make sure that others see the preview too. Thanks for checking.
