### 2024-08-05 #6967 @danielskatz
It seems like the issue where you didn't get the right version initially as it was being generated, and the bad version got cached (in the network upstream of any particular browser). If you have access to another network (e.g., perhaps a phone network), you might try it there. Otherwise, it should be refreshed in a couple of hours.
